@@ -5,9 +5,9 @@ class RatesController < ApplicationController
     rate = Rate.find_or_initialize_by(user: current_user, album_id: params[:album_id])
     rate.rate = params[:rate]
     if rate.save
-      render nothing: true, status: 201
+      head :created
     else
-      render nothing: true, status: 400
+      head :bad_request
     end
   end
 end
